requirements.txt
aiohttp==3.8.1
pymongo[srv]
motor==2.5.1
Pyrogram==2.0.30
TgCrypto==1.2.3
bs4
pyshorteners
aiofiles
dnspythons
python-dotenv
errorsolvers
domainers 
